Manually link to composer-bin-plugin installed commands
Otherwise there is no guarantee which version of the library the link was created for, as there might be still multiple versions installed. The plugin will link to the first one that was installed. It might not be the one we expected.

i.e. rector installed phpstan in version 0.11.3 as one of its dependencies. phpstan 0.11.4 was also directly requested to be installed in a different composer-bin-plugin namespace. Since rectors phpstan was installed first, the link was pointing to the 0.11.3 version instead of 0.11.4 we expected. As a consequence no phpstan plugins could be loaded.